js/xpconnect/tests/chrome/test_bug616992.xhtml

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=616992
-->
<window title="Mozilla Bug 601803"
  xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">
  <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=616992"
     target="_blank">Mozilla Bug 616992</a>

  <!-- test code goes here -->
  <script type="application/javascript"><![CDATA[

  /** Test for Bug 616992 **/

  var sandbox = new Cu.Sandbox(window);
  sandbox.w = window;
  var actual = Cu.evalInSandbox("Object.keys(w.NodeFilter)", sandbox).sort().toString();
  var expected = Object.keys(NodeFilter).sort().toString();
  is(actual, expected, "interface constants are visible inside sandboxes");

  ]]></script>
  </body>
</window>
